refactor: fix flex box arrangement to accomodate delete component
