--- frontend/components/account/SettingDelete.tsx
@@ -1,11 +1,0 @@
-function SettingDelete() {
-    return (
-        <>
-            <div className="flex flex-grow w- h-[calc(100vh-527px)] z-10">
-                <span className="h-1 w-full border-solid bg-slate-200"></span>
-            </div>
-        </>
-    )
-}
-
-export default SettingDelete
